[RAPTOR] Minimizer: Prefer right-most candidate
Right-most candidate means that if there are multiple k-mers with the
minimal hash value, I choose the one that is furthest to the right.

Example: A sequence with 48 A. window=23, kmer=19
After 23 characters, the first minimiser must be reported.
After that, there needs to be a new minimiser after every 5 characters
(there are 23 - 19 + 1 = 5 kmers in each window).
1 minimiser after 23 characters
48 - 23 = 25 characters left
25 / 5 = 5 5 minimisers for 25 characters
So there is a total of 6 minimisers, when the right-most candidate is
prefered.

When choosing the left-most candidate, I have a new minimiser each time
the window is moved.
So there would be
1 minimiser after 23 characters
48 - 23 = 25 characters left
25 / 1 = 25 minimiser for 25 characters
A total of 26 minimiser.

The same happens for palindromic (forward and reverse complement are the
same) sequences. The palindromic sequence might need to be > k.
At first, the forward strand is chosen.
When this one leaves the window, the reverse complement is chosen.
This leads to reporting the same minimiser value twice, which would not
happen when favoring the right-most candidate.